refactor(operator_commands_dashboard/routes): extract inline tests to siblings
Split inline tests at end of routes.rs (749 lines) into two sibling files
to begin bringing the 6395-LOC behemoth under the 400-LOC cap (#1266).

- routes.rs: 6395 -> 5632
- tests_routes_a.rs (NEW, 392): host helpers, hosts API tests
- tests_routes_b.rs (NEW, 365): tmux parsing, ws/distributed helpers

Promoted to pub(crate) to enable sibling test imports:
build_agent_graph, classify_agent_layer, format_recent_actions_for_cycle,
host_entry_name, is_local_host, load_hosts, read_recent_cycle_reports,
remote_vms_from_hosts, run_gh_json, tag_local_membership,
truncate_with_ellipsis (renamed from local truncate_outcome_detail),
LOGIN_HTML.

First of N PRs to bring routes.rs under 400 LOC.

Refs #1266
